+++ src/app/violation/page.tsx
@@ -1,15 +1,15 @@
 'use client'
-import { HomeTemplate } from 'components/Home/template/style';
-import Sidebar from 'components/common/organisms/Sidebar';
-import { CommonTemplate } from 'components/common/templates/style';
+import { HomeTemplate } from 'components/Home/template/style'
+import { ViolationTemplate } from 'components/Violation/templates/style'
+import Sidebar from 'components/common/organisms/Sidebar'
+import { CommonTemplate } from 'components/common/templates/style'
 
 const Violation = () => {
-    return (
-        <CommonTemplate>
-          <Sidebar />
-          <HomeTemplate>
-          </HomeTemplate>
-        </CommonTemplate>
-      );
+  return (
+    <CommonTemplate>
+      <Sidebar />
+      <ViolationTemplate></ViolationTemplate>
+    </CommonTemplate>
+  )
 }
 export default Violation
