+++ src/components/Product/organisms/style.ts
@@ -1,5 +1,4 @@
 import styled from '@emotion/styled'
-import { Palette } from 'style/global'
 
 export const Wrapper = styled.div``
 
@@ -13,10 +12,3 @@ export const BtnBox = styled.div`
   margin-top: 60px;
 `
 export const DeleteWrapper = styled.div``
-
-export const LineWrapper = styled.div`
-  width: 100%;
-  height: 2px;
-  background-color: #c2c2c2;
-  margin-bottom: 0.5rem;
-`
